Made the inverse function faster for 4x4 matrices. This should be roughly 100% faster for floats and 60% faster for duals than the current function.

```julia
using StaticArrays
using LinearAlgebra
using ForwardDiff
using BenchmarkTools

estOverhead(A::SMatrix{4,4,T,16}) where {T} = nothing

### Before
sm44 = randn(SMatrix{4,4,Float64,16})
@btime inv(sm44)          #  46.160 ns (1 allocation: 144 bytes)
@btime estOverhead(sm44)  #   7.740 ns (0 allocations: 0 bytes)

sm44 = randn(SMatrix{4,4,ForwardDiff.Dual{Nothing,Float64,6},16})
@btime inv(sm44)          # 357.433 ns (1 allocation: 1008 bytes)
@btime estOverhead(sm44)  #   8.376 ns (0 allocations: 0 bytes)

### After 
sm44 = randn(SMatrix{4,4,Float64,16})
@btime inv(sm44)          #  25.904 ns (1 allocation: 144 bytes)
@btime estOverhead(sm44)  #   7.740 ns (0 allocations: 0 bytes)

sm44 = randn(SMatrix{4,4,ForwardDiff.Dual{Nothing,Float64,6},16})
@btime inv(sm44)          # 220.067 ns (1 allocation: 1008 bytes)
@btime estOverhead(sm44)  #   8.376 ns (0 allocations: 0 bytes)
```