@KristofferC 2019-03-18
Try run the benchmarks with interpolating 

```
@btime inv($sm44)
```

etc.