@ryanelandt 2019-03-18
As does inv...

```julia

sm44 = randn(SMatrix{4,4,Float64,16})
@btime inv($sm44)          #   0.014 ns (0 allocations: 0 bytes)
```