@KristofferC 2019-03-18
Oh, sometimes I use

```jl
@btime inv($(Ref(sm44))[])
```

which usually works around it.