@ryanelandt 2019-03-18
Thanks I didn't realize this. Here is a corrected test:

```julia
using StaticArrays
using LinearAlgebra
using ForwardDiff
using BenchmarkTools

### Before
sm44 = randn(SMatrix{4,4,Float64,16})
@btime inv($(Ref(sm44))[])     #  60.706 ns (0 allocations: 0 bytes)

sm44 = randn(SMatrix{4,4,ForwardDiff.Dual{Nothing,Float64,6},16})
@btime inv($(Ref(sm44))[])     #  518.232 ns (0 allocations: 0 bytes)

### After
sm44 = randn(SMatrix{4,4,Float64,16})
@btime inv($(Ref(sm44))[])  #  28.310 ns (0 allocations: 0 bytes)

sm44 = randn(SMatrix{4,4,ForwardDiff.Dual{Nothing,Float64,6},16})
@btime inv($(Ref(sm44))[])  #  346.906 ns (0 allocations: 0 bytes)
```