@ryanelandt 2019-03-19
I did a test that compared the accuracy of this method and the old method. I do not know the right way to test two versions of a function so I copied both versions into functions called ```invert``` and ```invert_old``` and did a test this way. The new function has roughly 10% less error for random matrices and 25% less error for positive matrices. Both solutions are essentially identical, the old one is an expansion of the new one. My guess is that as the new one involves less floating point operations it induces less error.

```julia
using StaticArrays
using LinearAlgebra

function eval_err(is_make_pos::Bool)
	cum_err_old = 0.0
	cum_err_new = 0.0
	for k = 1:1e6
		sm44 = randn(SMatrix{4,4,Float64,16})
		is_make_pos && (sm44 = abs.(sm44))
		cum_err_old += norm(sm44 * invert_old(sm44) - I)
		cum_err_new += norm(sm44 * invert(sm44)     - I)
	end
	println("")
	println("cum_err_old: ", cum_err_old)
	println("cum_err_new: ", cum_err_new)
	println("new method is $(100 * (1 - cum_err_new / cum_err_old)) % more accurate than old method")
end


eval_err(false)
# cum_err_old: 5.442016380124342e-9
# cum_err_new: 4.801225758445484e-9
# new method is 11.774874916201883 % more accurate than old method

eval_err(true)
# cum_err_old: 1.4806311491135073e-8
# cum_err_new: 1.0418722581347007e-8
# new method is 29.633233857163077 % more accurate than old method
```