@c42f 2019-03-19
Using the additional stress test in `test/inv.jl` (https://github.com/JuliaArrays/StaticArrays.jl/blob/master/test/inv.jl#L102) shows this is an improvement for "almost rank 1" 4x4 matrices. For other approximately low rank cases it fares similarly to the code on master.

master:
![master](https://user-images.githubusercontent.com/601473/54581540-e3be7c80-4a58-11e9-98ae-783f571c87ec.png)

this pr:
![pr_597](https://user-images.githubusercontent.com/601473/54581591-110b2a80-4a59-11e9-9656-d2f39267de05.png)