@c42f 2019-03-19
Yes, thanks @ryanelandt.

It's interesting to note that the low rank numerical problems go away for the 5x5 case because we use the static LU decomposition there rather than special case code.

Similarly, if we comment out the 4x4 special case entirely, we have vastly improved numerical stability due to using LU, but there's a 7x speed hit to pay on my machine for that extra accuracy. Yikes!

```julia
julia> sm44 = randn(SMatrix{4,4,Float64,16})

# master, without 4x4 special case
julia> @btime inv($(Ref(sm44))[])
  136.877 ns (0 allocations: 0 bytes)

# ...
# master after merging this PR
julia> @btime inv($(Ref(sm44))[])
  18.941 ns (0 allocations: 0 bytes)
```